pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "movie-agent-service"
version = "0.1.2"
description = "Production-grade core library for AI-powered movie discovery and analysis using RAG and tool-calling agent architecture"
authors = [
    {name = "Alireza Barzin Zanganeh", email = ""}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["ai", "agent", "llm", "rag", "movie", "recommendation", "langchain", "faiss"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core LangChain Framework
    "langchain>=0.1.0,<1",
    "langchain-community==0.0.38",
    "langchain-core==0.1.52",
    "langchain-groq==0.1.4",
    "langchain-openai==0.1.6",
    "langchain-text-splitters==0.0.2",
    # Vector Store
    "faiss-cpu",
    # LLM Providers
    "groq",
    # Web Framework
    "flask>=2.3.0",
    "python-dotenv>=1.0.0",
    # Data Processing
    "pandas>=2.0.0",
    "numpy<2",
    # Text Processing & Fuzzy Matching
    "rapidfuzz>=3.0.0",
    "tiktoken>=0.5.0",
    # Image Processing
    "Pillow>=10.0.0",
    # Vision Models (optional - for local BLIP)
    "transformers>=4.35,<5",
    "torch==2.2.2",
    # Async Support
    "nest-asyncio>=1.5.0",
    # Utilities
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "black",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/abzanganeh/movie-agent-service"
Repository = "https://github.com/abzanganeh/movie-agent-service"
Documentation = "https://www.zanganehai.com/projects/movie-agent-service"
"Bug Tracker" = "https://github.com/abzanganeh/movie-agent-service/issues"

[tool.setuptools]
packages = ["movie_agent"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

